refactor: code quality improvements
- Replace assert_eq! in audio callback with graceful error handling
  to prevent DAW crashes on channel mismatch (outputs silence instead)

- Add 21 unit tests to voidmic_core (previously 0):
  NoiseFloorTracker (5), ThreeBandEq (3), LookaheadLimiter (4),
  VoidProcessor (6), FrameAdapter (3)

- Extract FrameAdapter to voidmic_core to deduplicate ring buffer
  interleave/process/deinterleave pattern shared by plugin and LV2

- Split gui.rs (1424 lines) into 8 focused modules:
  app.rs, engine.rs, tray.rs, devices.rs, controls.rs,
  advanced.rs, wizard.rs, mod.rs

- Fix pre-existing clippy warnings (too_many_arguments, collapsible_else_if)
