+++ crates/app/src/gui/mod.rs
@@ -1,0 +1,11 @@
+//! VoidMic GUI — modular implementation.
+
+mod advanced;
+mod app;
+mod controls;
+mod devices;
+mod engine;
+mod tray;
+mod wizard;
+
+pub use app::run_gui;
