+++ crates/core/Cargo.toml
@@ -12,3 +12,4 @@ aec3 = "0.1.3"
 anyhow = "1.0"
 log = "0.4"
 crossbeam-channel = "0.5.15"
+ringbuf = "0.4.7"
