+++ crates/core/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod constants;
 pub mod echo_cancel;
+pub mod frame_adapter;
 pub mod processor;
 
+pub use frame_adapter::FrameAdapter;
 pub use nnnoiseless::DenoiseState;
 pub use processor::VoidProcessor;
