+++ crates/lv2/Cargo.toml
@@ -9,6 +9,5 @@ crate-type = ["cdylib"]
 [dependencies]
 lv2 = "0.6"
 voidmic_core = { path = "../core" }
-ringbuf = "0.4.7"
 log = "0.4"
