+++ crates/plugin/Cargo.toml
@@ -9,7 +9,6 @@ crate-type = ["cdylib"]
 [dependencies]
 voidmic_core = { path = "../core" }
 nih_plug = { git = "https://github.com/robbert-vdh/nih-plug.git" }
-ringbuf = "0.4.7"
 anyhow = "1.0"
 nih_plug_egui = { git = "https://github.com/robbert-vdh/nih-plug.git" }
 egui = "0.31"
