+++ crates/plugin/src/lib.rs
@@ -1,25 +1,19 @@
 use crossbeam_channel::Receiver;
 use nih_plug::prelude::*;
 use nih_plug_egui::{create_egui_editor, widgets, EguiState};
-use ringbuf::traits::{Consumer, Observer, Producer};
-use ringbuf::HeapRb;
 use std::num::NonZeroU32;
 use std::sync::atomic::{AtomicU32, Ordering};
 use std::sync::Arc;
-use voidmic_core::constants::{FRAME_SIZE, SAMPLE_RATE};
-use voidmic_core::VoidProcessor;
+use voidmic_core::constants::SAMPLE_RATE;
+use voidmic_core::{FrameAdapter, VoidProcessor};
 use voidmic_ui::{theme, visualizer, widgets as ui_widgets};
 
 struct VoidMicPlugin {
     params: Arc<VoidMicParams>,
-    // editor_state removed as it is in params
 
     // Audio Processing State
     processor: Option<VoidProcessor>,
-
-    // Ring Buffers (Audio I/O)
-    rb_in: Option<HeapRb<f32>>,
-    rb_out: Option<HeapRb<f32>>,
+    adapter: Option<FrameAdapter>,
 
     // GUI Data Bridging
     volume_level: Arc<AtomicU32>,
@@ -56,8 +50,7 @@ impl Default for VoidMicPlugin {
         Self {
             params: Arc::new(VoidMicParams::default()),
             processor: None,
-            rb_in: None,
-            rb_out: None,
+            adapter: None,
             volume_level: Arc::new(AtomicU32::new(0)),
             spectrum_receiver: None,
         }
@@ -218,12 +211,7 @@ impl Plugin for VoidMicPlugin {
 
         self.volume_level = processor.volume_level.clone();
         self.processor = Some(processor);
-
-        let buffer_size = FRAME_SIZE * 4 * 2; // Always stereo
-
-        // Ringbuf 0.4
-        self.rb_in = Some(HeapRb::<f32>::new(buffer_size));
-        self.rb_out = Some(HeapRb::<f32>::new(buffer_size));
+        self.adapter = Some(FrameAdapter::new());
 
         true
     }
@@ -236,6 +224,10 @@ impl Plugin for VoidMicPlugin {
     ) -> ProcessStatus {
         let processor = match self.processor.as_mut() {
             Some(p) => p,
+            None => return ProcessStatus::Normal,
+        };
+        let adapter = match self.adapter.as_mut() {
+            Some(a) => a,
             None => return ProcessStatus::Normal,
         };
 
@@ -255,66 +247,31 @@ impl Plugin for VoidMicPlugin {
         }
         let num_samples = channel_data[0].len();
 
-        // 1. Push Input (Interleaved)
-        if let Some(rb_in) = &mut self.rb_in {
-            for i in 0..num_samples {
-                if num_channels == 2 {
-                    let _ = rb_in.try_push(channel_data[0][i]);
-                    let _ = rb_in.try_push(channel_data[1][i]);
-                } else if num_channels == 1 {
-                    // Duplicate Mono to Stereo
-                    let val = channel_data[0][i];
-                    let _ = rb_in.try_push(val);
-                    let _ = rb_in.try_push(val);
-                }
-            }
+        // 1. Push Input
+        if num_channels == 2 {
+            adapter.push_stereo_interleaved(&channel_data[0][..num_samples], &channel_data[1][..num_samples]);
+        } else if num_channels == 1 {
+            adapter.push_mono(&channel_data[0][..num_samples]);
         }
 
-        // 2. Process chunks
-        if let (Some(rb_in), Some(rb_out)) = (&mut self.rb_in, &mut self.rb_out) {
-            let mut left_in = [0.0f32; FRAME_SIZE];
-            let mut right_in = [0.0f32; FRAME_SIZE];
-            let mut left_out = [0.0f32; FRAME_SIZE];
-            let mut right_out = [0.0f32; FRAME_SIZE];
-
-            // Need 2 * FRAME_SIZE samples for a full stereo frame
-            while rb_in.occupied_len() >= FRAME_SIZE * 2 {
-                for j in 0..FRAME_SIZE {
-                    left_in[j] = rb_in.try_pop().unwrap_or(0.0);
-                    right_in[j] = rb_in.try_pop().unwrap_or(0.0);
-                }
-
-                processor.process_frame(
-                    &[&left_in, &right_in],
-                    &mut [&mut left_out, &mut right_out],
-                    None,
-                    self.params.suppression.value(),
-                    self.params.gate_threshold.value(),
-                    true,
-                );
-
-                for j in 0..FRAME_SIZE {
-                    let _ = rb_out.try_push(left_out[j]);
-                    let _ = rb_out.try_push(right_out[j]);
-                }
-            }
-        }
+        // 2. Process available frames
+        adapter.process_available(
+            processor,
+            self.params.suppression.value(),
+            self.params.gate_threshold.value(),
+            true,
+        );
 
         // 3. Output
-        if let Some(rb_out) = &mut self.rb_out {
-            for i in 0..num_samples {
-                if rb_out.occupied_len() >= 2 {
-                    let l = rb_out.try_pop().unwrap_or(0.0);
-                    let r = rb_out.try_pop().unwrap_or(0.0);
-
-                    if num_channels == 1 {
-                        channel_data[0][i] = (l + r) * 0.5;
-                    } else {
-                        channel_data[0][i] = l;
-                        channel_data[1][i] = r;
-                    }
-                }
-            }
+        if num_channels == 1 {
+            adapter.pop_mono(&mut channel_data[0][..num_samples]);
+        } else {
+            // Split borrows: we need mutable references to two different slices
+            let (left_slice, rest) = channel_data.split_at_mut(1);
+            adapter.pop_stereo(
+                &mut left_slice[0][..num_samples],
+                &mut rest[0][..num_samples],
+            );
         }
 
         ProcessStatus::Normal
